+++ src/engine/Util/EasingFunctions.ts
@@ -1,4 +1,6 @@
-﻿/**
+﻿import { Vector } from '../Algebra';
+
+/**
  * A definition of an EasingFunction. See [[EasingFunctions]].    
  */
 export interface EasingFunction { // tslint:disable-line
@@ -46,26 +48,48 @@ export interface EasingFunction { // tslint:disable-line
  * ```
  */
 export class EasingFunctions {
+   public static CreateReversableEasingFunction(easing: EasingFunction) {
+      return (time: number, start: number, end: number, duration: number) => {
+         if (end < start) {
+            return start - (easing(time, end, start, duration) - end);
+         } else {
+            return easing(time, start, end, duration);
+         }
+      };
+   }
+   
+   public static CreateVectorEasingFunction(easing: EasingFunction) {
+      return (time: number, start: Vector, end: Vector, duration: number) => {
+         return new Vector(
+            easing(time, start.x, end.x, duration),
+            easing(time, start.y, end.y, duration)
+         );
+      };
+   }
 
-   public static Linear: EasingFunction = (currentTime: number, startValue: number, endValue: number, duration: number)  => {
+   public static Linear: EasingFunction = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number)  => {
       endValue = (endValue - startValue);
       return endValue * currentTime / duration + startValue;
-   };
+   });
 
-   public static EaseInQuad = (currentTime: number, startValue: number, endValue: number, duration: number) => {
+   public static EaseInQuad = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number) => {
       endValue = (endValue - startValue);
       currentTime /= duration;
 
       return endValue * currentTime * currentTime + startValue;
-   };
+   });
    
-   public static EaseOutQuad: EasingFunction = (currentTime: number, startValue: number, endValue: number, duration: number) => {
+   public static EaseOutQuad: EasingFunction = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number) => {
       endValue = (endValue - startValue);
       currentTime /= duration;
       return -endValue * currentTime * (currentTime - 2) + startValue;
-   };
+   });
 
-   public static EaseInOutQuad: EasingFunction = (currentTime: number, startValue: number, endValue: number, duration: number) => {
+   public static EaseInOutQuad: EasingFunction = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number) => {
       endValue = (endValue - startValue);
       currentTime /= duration / 2;
 
@@ -73,27 +97,30 @@ export class EasingFunctions {
       currentTime--;
 
       return -endValue / 2 * (currentTime * (currentTime - 2) - 1) + startValue;
-   };
+   });
 
-   public static EaseInCubic: EasingFunction = (currentTime: number, startValue: number, endValue: number, duration: number) => {
+   public static EaseInCubic: EasingFunction = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number) => {
       endValue = (endValue - startValue);
       currentTime /= duration;
       return endValue * currentTime * currentTime * currentTime + startValue;
-   };
+   });
 
-   public static EaseOutCubic: EasingFunction = (currentTime: number, startValue: number, endValue: number, duration: number) => {
+   public static EaseOutCubic: EasingFunction = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number) => {
       endValue = (endValue - startValue);
       currentTime /= duration;
       currentTime--;
       return endValue * (currentTime * currentTime * currentTime + 1) + startValue;
-   };
+   });
 
-   public static EaseInOutCubic: EasingFunction = (currentTime: number, startValue: number, endValue: number, duration: number) => {
+   public static EaseInOutCubic: EasingFunction = EasingFunctions.CreateReversableEasingFunction(
+      (currentTime: number, startValue: number, endValue: number, duration: number) => {
       endValue = (endValue - startValue);
       currentTime /= duration / 2;
       if (currentTime < 1) { return endValue / 2 * currentTime * currentTime * currentTime + startValue; }
       currentTime -= 2;
       return endValue / 2 * (currentTime * currentTime * currentTime + 2) + startValue;
-   };
+   });
    
 }